+++ acrg_obs/read.py
@@ -789,9 +789,8 @@ def get_obs(sites, species,
     network = check_list_and_length(network, sites, "network")
     instrument = check_list_and_length(instrument, sites, "instrument")
 
-    if data_directory is not None:
-        # Need to create a database in data_directory
-        obs_database(data_directory = data_directory)
+    #Need to create a database in data_directory
+    obs_database(data_directory = data_directory)
     
     # Get data
     obs = {}